+++ src/domain/types/window.d.ts
