+++ src/main.tsx
@@ -1,7 +1,7 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
-import App from './App.tsx';
-import './index.css';
+import App from '@presentation/App.tsx';
+import '@presentation/index.css';
 
 createRoot(document.getElementById('root')!).render(
   <StrictMode>
