+++ src/presentation/components/AdvancedSettingsModal.tsx
