+++ src/presentation/components/WalletConnectModal.tsx
@@ -3,9 +3,9 @@ import { X, Wallet } from "lucide-react";
 import { Button } from "./ui/button";
 import { Card, CardContent, CardHeader } from "./ui/card";
 import { useConnect } from "wagmi";
-import braveLogo from "../assets/wallets-icons/brave-logo.png";
-import metamask from "../assets/wallets-icons/metamask.svg";
-import coibaseLogo from "../assets/wallets-icons/coinbase.svg";
+import braveLogo from "@assets/wallets-icons/brave-logo.png";
+import metamask from "@assets/wallets-icons/metamask.svg";
+import coibaseLogo from "@assets/wallets-icons/coinbase.svg";
 
 interface WalletConnectModalProps {
   isOpen: boolean;
