+++ src/presentation/components/ui/dialog.tsx
@@ -2,7 +2,7 @@ import * as React from "react";
 import * as DialogPrimitive from "@radix-ui/react-dialog";
 import { X } from "lucide-react";
 
-import { cn } from "../../lib/utils";
+import { cn } from "@shared/lib/utils";
 
 const Dialog = DialogPrimitive.Root;
