+++ vite.config.ts
@@ -1,9 +1,19 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
+import path from "node:path";
 
 // https://vitejs.dev/config/
 export default defineConfig({
   plugins: [react()],
+  resolve: {
+    alias: {
+      "@application": path.resolve(__dirname, "src/application"),
+      "@domain": path.resolve(__dirname, "src/domain"),
+      "@presentation": path.resolve(__dirname, "src/presentation"),
+      "@shared": path.resolve(__dirname, "src/shared"),
+      "@assets": path.resolve(__dirname, "src/assets"),
+    },
+  },
   optimizeDeps: {
     exclude: ["lucide-react"],
   },
